zoom functionality added to map via answer select
