+++ client/src/containers/QuizContainer.js
@@ -1,7 +1,8 @@
 import { useState } from "react";
 import QuestionComponent from '../components/QuestionComponent';
 import ResultComponent from '../components/ResultComponent';
-import MapComponent from '../components/MapComponent';
+import { MapContainer, TileLayer, Marker, Popup, useMap } from 'react-leaflet';
+import L from 'leaflet';
 
 const QuizContainer = ({ selectedQuestions, selectedCountry }) => {
 
@@ -29,21 +30,47 @@ const QuizContainer = ({ selectedQuestions, selectedCountry }) => {
     }
   };
 
-  if (questionCount !== 4) {
-    return (
-      <>
-        <MapComponent country={selectedCountry} currentZoomLevel={questionCount}/>
-        <QuestionComponent quizQuestions={selectedQuestions} currentQuestion={questionCount} increaseQuestionCount={increaseQuestionCount} checkAnswerCorrect={checkAnswerCorrect} />
-      </>
-    )
-  } else {
-    return (
-      <>
-        <MapComponent country={selectedCountry} currentZoomLevel={questionCount}/>
-        <ResultComponent score={correctAnswerCount}/>
-      </>
-    )
+  const markerIcon = L.icon({
+    // some generic icon from google
+    iconUrl: "https://cdn0.iconfinder.com/data/icons/small-n-flat/24/678111-map-marker-512.png",
+    iconSize: [35, 35]
+  });
+
+
+  const resultToggle = () => {
+    if (questionCount !== 4) {
+      return <QuestionComponent quizQuestions={selectedQuestions} currentQuestion={questionCount} increaseQuestionCount={increaseQuestionCount} checkAnswerCorrect={checkAnswerCorrect} />
+    } else {
+      return <ResultComponent score={correctAnswerCount} />
+    }
   }
+
+  const SetViewOnClick = ({zoom}) => {
+    const map = useMap();
+    console.log("set zoom")
+    map.setZoom(zoom);
+
+    return null;
+  }
+
+  return (
+    <>
+      <MapContainer center={selectedCountry.lat_long} zoom={selectedCountry.zooms[questionCount]} scrollWheelZoom={false}>
+        <TileLayer
+          attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
+          url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+        />
+        <Marker position={selectedCountry.lat_long} icon={markerIcon}>
+          <Popup>
+            A pretty CSS3 popup. <br /> Easily customizable.
+        </Popup>
+        </Marker>
+        <SetViewOnClick zoom={selectedCountry.zooms[questionCount]} />
+      </MapContainer>
+      {resultToggle()}
+
+    </>
+  )
 
 };
